Bulk replace a pattern throughout multiple files.

This function makes it easier to change a specific text string throughout a number of files, allowing you to ensure you're correcting all areas of your code at once.

bulk_replace(files, pattern, replacement, dry.run = TRUE)

Arguments

files
A vector of filepaths to replace strings in.

pattern
The character string to be replaced.

replacement
A character string to replace all patterns with.

dry.run
Logical -- describe the file changes that will be made (TRUE) or make them in the specified files (FALSE)?

Examples

library(heddlr)
temp_file <- tempfile("test", fileext = ".Rmd")
temp_patt <- "#"
export_template(temp_patt, temp_file)
bulk_replace(c(temp_file), "#", "##")

#> Running this for real would have changed 1 characters across 1 files. (Set dry.run = F to make these changes for real.)

bulk_replace(c(temp_file), "#", "##", dry.run = FALSE)