Safely export templates to file.

Safely export templates to file.

export_template(template, filename, sep = "",
  filename.is.string = TRUE)

Arguments

template
The template string to be written out

filename
The path to write the template to, passed to writeLines. Also accepts stdout (and likely other similar functions) with a warning.

sep
Separator to use between lines written, passed to writeLines. Defaults to no separator, as templates are generally already spaced appropriately.

filename.is.string
A binary value indicating whether or not the filename parameter is expected to be a string (that is, a character vector). Setting the value to FALSE disables the warning when a non-character argument is passed, but this is unsupported functionality.

Details

Note that this function is effectively the inverse of import_pattern -- export_template(import_pattern("out.txt"), "out.txt") should always result in an unchanged file, and exceptions to this rule would be considered bugs.