Easily provide parameters to R Markdown render calls

R Markdown documents allow you to pass almost any object -- including large data frames and functions -- to the document as parameters, letting you only define them once to use them in both your document generator and the generated document. This function makes it slightly easier to do so, by automatically creating a named list from provided objects rather than requiring a named list. This function is a stripped-down variant of lst.

provide_parameters(...)

Arguments

...
Objects to be included as parameters. Objects should be unquoted and exist in the current session environment.

See also

Examples

template <- make_template(
  "---\ntitle: Example\noutput: html_document\n---\n",
  "\nThe random number is `r random_number`.\n"
)
template <- use_parameters(template, "random_number")
pattern_file <- tempfile("out", tempdir(), ".Rmd")
export_template(template, pattern_file)

random_number <- rnorm(1)
if (rmarkdown::pandoc_available()) {
  rmarkdown::render(pattern_file, params = provide_parameters(random_number))
}

#> 
#> 
#> processing file: out5056378524e1.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |.......................                                               |  33%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............................................                       |  67%
#> label: unnamed-chunk-1
#>   |                                                                              |......................................................................| 100%
#>    inline R code fragments
#> 
#> 

#> output file: out5056378524e1.knit.md

#> /home/mikemahoney218/anaconda3/bin/pandoc +RTS -K512m -RTS out5056378524e1.utf8.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash+smart --output out5056378524e1.html --email-obfuscation none --self-contained --standalone --section-divs --template /home/mikemahoney218/R/x86_64-pc-linux-gnu-library/3.6/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --variable 'theme:bootstrap' --include-in-header /tmp/RtmpQEx0Qk/rmarkdown-str5056449f0c85.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --lua-filter /home/mikemahoney218/R/x86_64-pc-linux-gnu-library/3.6/rmarkdown/rmd/lua/pagebreak.lua --lua-filter /home/mikemahoney218/R/x86_64-pc-linux-gnu-library/3.6/rmarkdown/rmd/lua/latex-div.lua 

#> 
#> Output created: out5056378524e1.html